Titan: Rename Add Mailboxes component with more accurate name
